multimailhook.dontComputeNewRevs: disable new commits detection on demand
This is useful as a workaround when using the script as an update hook.

I use git-multimail to track changes to origin/master after a fetch, and
the new commits detection goes wrong most of the time because the fetch
updated both origin/master and origin/HEAD, and the hook is ran after the
actual update. In this use-case, the new commit detection is useless and
counter-productive.