normalize changed ref names in computation of exclude revspec
When the user calls git-multimail manually, the refname may be given as a
short name (e.g. master), but the comparison done a few lines after the
change in _compute_other_ref_sha1s() to exclude refs that are part of the
current change is done on the full name (e.g. refs/heads/master).
Normalize the ref names before the comparison to fix this.